pyproject.toml
[project]
name = "frankfurtermcp"
version = "0.1.0"
description = "A MCP server for the Frankfurter API for currency exchange rates."
readme = "README.md"
license = "MIT"
authors = [
    { name = "Anirban Basu", email = "anirbanbasu@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=2.7.0",
    "python-dotenv>=1.1.0",
]

[project.scripts]
frankfurtermcp = "frankfurtermcp.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "icecream>=2.1.4",
]
test = [
    "pytest>=8.4.0",
]
